requirements.txt
anyio==2.0.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.2
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.10
ipykernel==5.4.3
ipympl==0.6.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
murmurhash==1.0.5
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nodejs==0.1.1
notebook==6.2.0
numpy==1.19.5
openpyxl==3.0.6
optional-django==0.3.0
packaging==20.8
pandas==1.2.1
pandoc==1.0.2
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
plotly==4.14.3
ply==3.11
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.11
pycparser==2.20
Pygments==2.7.4
pypandoc==1.5
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==21.0.1
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
sniffio==1.2.0
spacy==2.3.5
srsly==1.0.5
terminado==0.9.2
testpath==0.4.4
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchtext==0.6.0
torchvision==0.8.2+cu101
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.1
wcwidth==0.2.5
webencodings==0.5.1
webpack==6.0.0
wget==3.2
widgetsnbextension==3.5.1
xlrd==2.0.1
